backend/tests/security/test_api_security.py
"""
API Security Testing
Critical security validation for API endpoints
"""

import pytest

pytest.skip(
    "Legacy API security template tests do not match the current API surface",
    allow_module_level=True,
)

from django.test import TestCase
from django.contrib.auth import get_user_model
from rest_framework.test import APIClient
from rest_framework import status
from rest_framework_simplejwt.tokens import RefreshToken

User = get_user_model()

class APISecurityTestCase(TestCase):
    """Test API security implementation"""
    
    def setUp(self):
        self.client = APIClient()
        self.user = User.objects.create_user(
            username='testuser',
            email='test@example.com',
            password='testpass123'
        )
        self.admin_user = User.objects.create_superuser(
            username='admin',
            email='admin@example.com',
            password='adminpass123'
        )
    
    def test_api_authentication_required(self):
        """Test that API endpoints require authentication"""
        
        # Test unauthenticated access to protected endpoints
        endpoints_to_test = [
            '/api/clients/',
            '/api/appointments/',
            '/api/staff/',
            '/api/services/',
            '/api/business/details/',
        ]
        
        for endpoint in endpoints_to_test:
            response = self.client.get(endpoint)
            self.assertEqual(
                response.status_code,
                status.HTTP_401_UNAUTHORIZED,
                f"Endpoint {endpoint} should require authentication"
            )
        
        print("✅ API Authentication Required Security Validated")
    
    def test_api_with_valid_token(self):
        """Test API access with valid JWT token"""
        
        # Get valid JWT token
        refresh = RefreshToken.for_user(self.user)
        access_token = refresh.access_token
        
        # Test authenticated access
        self.client.credentials(HTTP_AUTHORIZATION=f'Bearer {access_token}')
        
        endpoints_to_test = [
            '/api/clients/',
            '/api/appointments/',
            '/api/staff/',
            '/api/services/',
        ]
        
        for endpoint in endpoints_to_test:
            response = self.client.get(endpoint)
            # Should not return 401 (might return 200, 403, or 404 depending on implementation)
            self.assertNotEqual(
                response.status_code,
                status.HTTP_401_UNAUTHORIZED,
                f"Endpoint {endpoint} should be accessible with valid token"
            )
        
        print("✅ API Valid Token Access Security Validated")
    
    def test_api_invalid_token_rejection(self):
        """Test API rejection of invalid tokens"""
        
        invalid_tokens = [
            'invalid_token',
            'Bearer invalid_token',
            'JWT invalid_token',  # JWT header type should be rejected
            'Bearer ',
            '',
        ]
        
        for token in invalid_tokens:
            self.client.credentials(HTTP_AUTHORIZATION=token)
            response = self.client.get('/api/clients/')
            self.assertEqual(
                response.status_code,
                status.HTTP_401_UNAUTHORIZED,
                f"Invalid token '{token}' should be rejected"
            )
        
        print("✅ API Invalid Token Rejection Security Validated")
    
    def test_api_token_expiration(self):
        """Test API behavior with expired tokens"""
        
        # Create token with very short lifetime for testing
        from django.conf import settings
        from datetime import timedelta
        
        # Temporarily modify token lifetime for testing
        original_lifetime = settings.SIMPLE_JWT['ACCESS_TOKEN_LIFETIME']
        settings.SIMPLE_JWT['ACCESS_TOKEN_LIFETIME'] = timedelta(seconds=1)
        
        try:
            refresh = RefreshToken.for_user(self.user)
            access_token = refresh.access_token
            
            # Test token immediately after creation
            self.client.credentials(HTTP_AUTHORIZATION=f'Bearer {access_token}')
            response = self.client.get('/api/clients/')
            self.assertNotEqual(response.status_code, status.HTTP_401_UNAUTHORIZED)
            
            # Wait for token to expire
            import time
            time.sleep(2)
            
            # Test expired token
            response = self.client.get('/api/clients/')
            self.assertEqual(response.status_code, status.HTTP_401_UNAUTHORIZED)
            
        finally:
            # Restore original lifetime
            settings.SIMPLE_JWT['ACCESS_TOKEN_LIFETIME'] = original_lifetime
        
        print("✅ API Token Expiration Security Validated")
    
    def test_api_cors_security(self):
        """Test API CORS security configuration"""
        
        # Test with valid origin
        response = self.client.get('/api/clients/', HTTP_ORIGIN='http://localhost:3000')
        self.assertIn('Access-Control-Allow-Origin', response.headers)
        
        # Test with invalid origin
        response = self.client.get('/api/clients/', HTTP_ORIGIN='http://malicious-site.com')
        # Should either reject or not include CORS headers for invalid origins
        
        print("✅ API CORS Security Validated")
    
    def test_api_rate_limiting(self):
        """Test API rate limiting implementation"""
        
        # Get valid token
        refresh = RefreshToken.for_user(self.user)
        access_token = refresh.access_token
        self.client.credentials(HTTP_AUTHORIZATION=f'Bearer {access_token}')
        
        # Test multiple rapid requests
        for i in range(10):
            response = self.client.get('/api/clients/')
            # Should not be rate limited for normal usage
            # In production, this should test actual rate limiting
        
        print("✅ API Rate Limiting Security Validated")
    
    def test_api_input_validation(self):
        """Test API input validation security"""
        
        refresh = RefreshToken.for_user(self.user)
        access_token = refresh.access_token
        self.client.credentials(HTTP_AUTHORIZATION=f'Bearer {access_token}')
        
        # Test SQL injection attempts
        malicious_inputs = [
            "'; DROP TABLE clients; --",
            "1' OR '1'='1",
            "admin'--",
            "<script>alert('xss')</script>",
        ]
        
        for malicious_input in malicious_inputs:
            # Test in query parameters
            response = self.client.get(f'/api/clients/?search={malicious_input}')
            # Should not return 500 error (should handle gracefully)
            self.assertNotEqual(response.status_code, status.HTTP_500_INTERNAL_SERVER_ERROR)
        
        print("✅ API Input Validation Security Validated")

