backend/requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djangorestframework-camel-case==1.4.2
django-filter==23.3
django-cors-headers==4.3.0

# Database
psycopg2-binary>=2.9.9  # Updated for Python 3.13 compatibility (2.9.11+ has pre-built wheels)

# Utilities
python-decouple==3.8
python-dateutil==2.8.2

# Excel/CSV Processing
pandas>=2.2.0  # Updated for Python 3.13 compatibility (uses pre-built wheels)
openpyxl==3.1.2

# PDF Generation
reportlab==4.0.7

# Async/WebSocket
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1
daphne==4.1.0  # ASGI server for WebSocket support and testing

# Background Tasks
celery==5.3.4

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-asyncio==0.25.2  # For async test support
factory-boy==3.3.3
SQLAlchemy==2.0.44  # Required for Python 3.13 compatibility with factory-boy

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Monitoring
sentry-sdk==1.38.0

# Development
ipython==8.17.2
django-debug-toolbar==4.2.0
